레포츠 평점 출력 수정

+++ src/main/java/com/dto/LeportsDTO.java
@@ -27,7 +27,7 @@ public class LeportsDTO {
 	private String leports_sub_img4;
 	private String leports_content;
 	private int review_cnt;
-	private int review_star;
+	private double review_star;
 	private int reserve_cnt;
 	
 }
